Prioritize main branch structure: - Remove custom gacha.dart and pointget.dart implementations - Use screens/ directory structure from main branch - Fix navigation references and integrate point system - Resolve all linter errors
