+++ lib/widgets/navigationbar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:takizawa_hackathon_vol8/pointget.dart';
+// import 'package:takizawa_hackathon_vol8/pointget.dart'; // 削除済み
 import 'package:takizawa_hackathon_vol8/screens/ranking.dart';
 import 'package:takizawa_hackathon_vol8/screens/profile.dart';
 import 'package:takizawa_hackathon_vol8/screens/gacha.dart';
@@ -97,11 +97,11 @@ class _MainNavigationScreenState extends ConsumerState<MainNavigationScreen> {
   Widget _buildCurrentScreen() {
     switch (_currentIndex) {
       case NavIndex.home:
-        return const ProfileScreen();
+        return const ProfileScreen(); // ホーム画面としてプロフィール画面を使用
       case NavIndex.ranking:
         return const RankingScreen();
       case NavIndex.profile:
-        return const PointGetScreen();
+        return const Center(child: Text('ポイント獲得画面\n（未実装）', style: TextStyle(fontSize: 18))); // プレースホルダー
       case NavIndex.menu:
         return const GachaScreen();
     }
